+++ src/createAMSClient.ts
@@ -1,12 +1,14 @@
 import AMSLogger from "./AMSLogger";
 import FramedClient from "./FramedClient";
 import FramedlessClient from "./FramedlessClient";
+import GlobalConfiguration from "./GlobalConfiguration";
 import PluggableLogger from "./PluggableLogger";
 
 interface AMSConfig {
     framedMode: boolean,
     debug?: boolean,
-    logger?: PluggableLogger
+    logger?: PluggableLogger,
+    silentError?: boolean
 }
 
 const createAMSClient = async (config: AMSConfig): Promise<FramedClient | FramedlessClient> => {
@@ -16,6 +18,9 @@ const createAMSClient = async (config: AMSConfig): Promise<FramedClient | Framed
     const logger = new AMSLogger(config.logger);
     const client = config.framedMode? new FramedClient(logger): new FramedlessClient(logger);
     (config as AMSConfig).debug && client.setDebug((config as AMSConfig).debug || false);
+
+    GlobalConfiguration.debug = (config as AMSConfig).debug || false;
+    GlobalConfiguration.silentError = (config as AMSConfig).silentError || true;
 
     return client;
 }